Deduplicate the flexbox/layout converters
