+++ redwood-layout-uiview/src/commonMain/kotlin/app/cash/redwood/layout/uiview/utils.kt
@@ -16,13 +16,11 @@
 package app.cash.redwood.layout.uiview
 
 import app.cash.redwood.LayoutModifier
-import app.cash.redwood.flexcontainer.AlignItems
 import app.cash.redwood.flexcontainer.AlignSelf
 import app.cash.redwood.flexcontainer.FlexDirection
 import app.cash.redwood.flexcontainer.FlexItem
 import app.cash.redwood.flexcontainer.FlexItem.Companion.DefaultFlexGrow
 import app.cash.redwood.flexcontainer.FlexItem.Companion.DefaultFlexShrink
-import app.cash.redwood.flexcontainer.JustifyContent
 import app.cash.redwood.flexcontainer.Measurable
 import app.cash.redwood.flexcontainer.MeasureSpec
 import app.cash.redwood.flexcontainer.MeasureSpecMode
@@ -36,7 +34,6 @@ import app.cash.redwood.layout.Padding as PaddingModifier
 import app.cash.redwood.layout.Shrink
 import app.cash.redwood.layout.VerticalAlignment
 import app.cash.redwood.layout.api.CrossAxisAlignment
-import app.cash.redwood.layout.api.MainAxisAlignment
 import app.cash.redwood.layout.api.Padding
 import kotlinx.cinterop.CValue
 import kotlinx.cinterop.useContents
@@ -47,32 +44,6 @@ import platform.UIKit.UIViewNoIntrinsicMetric
 import platform.UIKit.intrinsicContentSize
 import platform.UIKit.sizeThatFits
 import platform.UIKit.subviews
-
-internal fun MainAxisAlignment.toJustifyContent() = when (this) {
-  MainAxisAlignment.Start -> JustifyContent.FlexStart
-  MainAxisAlignment.Center -> JustifyContent.Center
-  MainAxisAlignment.End -> JustifyContent.FlexEnd
-  MainAxisAlignment.SpaceBetween -> JustifyContent.SpaceBetween
-  MainAxisAlignment.SpaceAround -> JustifyContent.SpaceAround
-  MainAxisAlignment.SpaceEvenly -> JustifyContent.SpaceEvenly
-  else -> throw AssertionError()
-}
-
-internal fun CrossAxisAlignment.toAlignItems() = when (this) {
-  CrossAxisAlignment.Start -> AlignItems.FlexStart
-  CrossAxisAlignment.Center -> AlignItems.Center
-  CrossAxisAlignment.End -> AlignItems.FlexEnd
-  CrossAxisAlignment.Stretch -> AlignItems.Stretch
-  else -> throw AssertionError()
-}
-
-internal fun CrossAxisAlignment.toAlignSelf() = when (this) {
-  CrossAxisAlignment.Start -> AlignSelf.FlexStart
-  CrossAxisAlignment.Center -> AlignSelf.Center
-  CrossAxisAlignment.End -> AlignSelf.FlexEnd
-  CrossAxisAlignment.Stretch -> AlignSelf.Stretch
-  else -> throw AssertionError()
-}
 
 internal fun Padding.toSpacing() = Spacing(start.toDouble(), end.toDouble(), top.toDouble(), bottom.toDouble())
